# perf: optimise `encodeRandom()` (10x improvement)