This change improves the performance of `encodeRandom()` and ULID generation by 10x.

Before:
```
cpu: Apple M2
runtime: deno 1.36.0 (aarch64-apple-darwin)

benchmark         time (avg)        iter/s             (min … max)       p75       p99      p995
------------------------------------------------------------------ -----------------------------
encodeTime       302.74 ns/iter   3,303,187.7  (298.18 ns … 334.8 ns) 303.59 ns 329.95 ns  334.8 ns
encodeRandom       5.83 µs/iter     171,409.6     (5.21 µs … 6.27 µs)   6.11 µs   6.27 µs   6.27 µs
generate           9.39 µs/iter     106,460.0     (9.08 µs … 9.68 µs)   9.57 µs   9.68 µs   9.68 µs
```

After:
```
cpu: Apple M2
runtime: deno 1.36.0 (aarch64-apple-darwin)
benchmark         time (avg)        iter/s             (min … max)       p75       p99      p995
------------------------------------------------------------------ -----------------------------
encodeTime       303.42 ns/iter   3,295,776.1  (299.3 ns … 330.98 ns) 304.59 ns 326.78 ns 330.98 ns
encodeRandom     577.64 ns/iter   1,731,186.6 (337.92 ns … 738.33 ns) 640.61 ns 738.33 ns 738.33 ns
generate         670.84 ns/iter   1,490,678.9 (612.39 ns … 721.08 ns) 694.94 ns 721.08 ns 721.08 ns
```